Make ZeissTIFFReader#isThisType() much more performant
It is inappropriate to invalidate the CaseInsensitiveLocation cache
every single time that the ZeissTIFFReader is run.

Additionally, this developer doubts that the implementation of the
insensitive location handling is sound. It should *at least* have a
look at the mtime of the directory (only evicting the cache if the
mtime has actually changed!). It should also *at least* try to find
the file in question with the given spelling, and only fall back to
enumerate the files in the same directory when the file was not found
with that spelling. It should never look into a directory whose path
displays different cases.

This single change brings the time of Tobias' benchmark down to 317ms
(from 957ms, no less).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>